ocms/requirements.txt
Django>=5.2
djangorestframework>=3.15
django-filter>=24.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
redis>=5.0
